+++ config/e2e/rc-e2e-manager.yaml
@@ -2,7 +2,7 @@ apiVersion: apps/v1
 kind: Deployment
 metadata:
   name: manager
-  namespace: system
+  namespace: e2e-operator
   labels:
     app.kubernetes.io/name: datadog-operator
     control-plane: controller-manager
