+++ config/new-e2e/kustomization.yaml
@@ -2,6 +2,11 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 namePrefix: datadog-operator-e2e-
 namespace: e2e-operator
+patches:
+- path: rc-e2e-manager.yaml
+  target:
+    kind: Deployment
+    name: manager
 resources:
 - ../crd
 - ../rbac
