+++ go.work
@@ -1,6 +1,6 @@
-go 1.22.7
+go 1.23.0
 
-toolchain go1.22.7
+toolchain go1.23.0
 
 use (
 	.
