+++ test/e2e/manifests/datadog-agent-logs.yaml
@@ -1,16 +1,12 @@
 apiVersion: datadoghq.com/v2alpha1
 kind: DatadogAgent
 metadata:
-  name: datadog
+  name: datadog-agent-logs
+  namespace: e2e-operator
+  labels:
+    agent.datadoghq.com/e2e-test: datadog-agent-logs
 spec:
   global:
-    credentials:
-      apiSecret:
-        keyName: api-key
-        secretName: datadog-secret
-      appSecret:
-        keyName: app-key
-        secretName: datadog-secret
     kubelet:
       tlsVerify: false
   features:
@@ -22,18 +18,3 @@ spec:
       containerCollectAll: true
     liveContainerCollection:
       enabled: true
-  override:
-    nodeAgent:
-      labels:
-        agent.datadoghq.com/e2e-test: datadog-agent-logs
-      containers:
-        agent:
-          env:
-            - name: DD_SKIP_SSL_VALIDATION
-              value: "true"
-      env:
-        - name: DD_CLUSTER_NAME
-          valueFrom:
-            configMapKeyRef:
-              name: datadog-cluster-name
-              key: DD_CLUSTER_NAME
