+++ test/e2e/provisioners/kindProvisioner.go
@@ -1,0 +1,152 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build e2e
+// +build e2e
+
+package provisioners
+
+import (
+	"fmt"
+	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/e2e"
+	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/environments"
+	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/runner"
+	"github.com/DataDog/datadog-operator/test/e2e/common"
+	"github.com/DataDog/test-infra-definitions/common/utils"
+	localKubernetes "github.com/DataDog/test-infra-definitions/components/kubernetes"
+	resAws "github.com/DataDog/test-infra-definitions/resources/aws"
+	"github.com/DataDog/test-infra-definitions/scenarios/aws/ec2"
+	"github.com/pulumi/pulumi-kubernetes/sdk/v4/go/kubernetes"
+	corev1 "github.com/pulumi/pulumi-kubernetes/sdk/v4/go/kubernetes/core/v1"
+	"github.com/pulumi/pulumi-kubernetes/sdk/v4/go/kubernetes/kustomize"
+	metav1 "github.com/pulumi/pulumi-kubernetes/sdk/v4/go/kubernetes/meta/v1"
+	"github.com/pulumi/pulumi/sdk/v3/go/auto"
+	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
+	"path/filepath"
+	"strings"
+)
+
+// kindProvisioner Pulumi E2E provisioner to deploy the Operator binary with kustomize and deploy DDA manifest
+func KindProvisioner(k8sVersion string, extraKustomizeResources []string) e2e.Provisioner {
+	return e2e.NewTypedPulumiProvisioner[environments.Kubernetes]("kind-operator", func(ctx *pulumi.Context, env *environments.Kubernetes) error {
+		// Provision AWS environment
+		awsEnv, err := resAws.NewEnvironment(ctx)
+		if err != nil {
+			return err
+		}
+
+		// Create EC2 VM
+		vm, err := ec2.NewVM(awsEnv, "kind", ec2.WithUserData(UserData))
+		if err != nil {
+			return err
+		}
+		if err := vm.Export(ctx, nil); err != nil {
+			return err
+		}
+
+		// Create kind cluster
+		kindClusterName := strings.ReplaceAll(ctx.Stack(), ".", "-")
+
+		err = ctx.Log.Info(fmt.Sprintf("Creating kind cluster with K8s version: %s", k8sVersion), nil)
+		if err != nil {
+			return err
+		}
+
+		installEcrCredsHelperCmd, err := ec2.InstallECRCredentialsHelper(awsEnv, vm)
+		if err != nil {
+			return err
+		}
+
+		kindCluster, err := localKubernetes.NewKindCluster(&awsEnv, vm, awsEnv.CommonNamer().ResourceName("kind"), k8sVersion, utils.PulumiDependsOn(installEcrCredsHelperCmd))
+		if err != nil {
+			return err
+		}
+		if err := kindCluster.Export(ctx, &env.KubernetesCluster.ClusterOutput); err != nil {
+			return err
+		}
+
+		// Build Kubernetes provider
+		kindKubeProvider, err := kubernetes.NewProvider(ctx, awsEnv.Namer.ResourceName("k8s-provider"), &kubernetes.ProviderArgs{
+			EnableServerSideApply: pulumi.BoolPtr(true),
+			Kubeconfig:            kindCluster.KubeConfig,
+		})
+		if err != nil {
+			return err
+		}
+
+		// Deploy resources from kustomize config/e2e directory
+		kustomizeDirPath, err := filepath.Abs(MgrKustomizeDirPath)
+		if err != nil {
+			return err
+		}
+
+		UpdateKustomization(kustomizeDirPath, extraKustomizeResources)
+
+		e2eKustomize, err := kustomize.NewDirectory(ctx, "e2e-manager",
+			kustomize.DirectoryArgs{
+				Directory: pulumi.String(kustomizeDirPath),
+			},
+			pulumi.Provider(kindKubeProvider))
+		if err != nil {
+			return err
+		}
+
+		pulumi.DependsOn([]pulumi.Resource{e2eKustomize})
+
+		// Create imagePullSecret to pull E2E operator image from ECR
+		if common.ImgPullPassword != "" {
+			_, err = utils.NewImagePullSecret(&awsEnv, common.NamespaceName, pulumi.Provider(kindKubeProvider))
+			if err != nil {
+				return err
+			}
+		}
+
+		// Create datadog agent secret
+		_, err = corev1.NewSecret(ctx, "datadog-secret", &corev1.SecretArgs{
+			Metadata: metav1.ObjectMetaArgs{
+				Namespace: pulumi.String(common.NamespaceName),
+				Name:      pulumi.String("datadog-secret"),
+			},
+			StringData: pulumi.StringMap{
+				"api-key": awsEnv.CommonEnvironment.AgentAPIKey(),
+				"app-key": awsEnv.CommonEnvironment.AgentAPPKey(),
+			},
+		}, pulumi.Provider(kindKubeProvider))
+		if err != nil {
+			return err
+		}
+
+		// Create datadog cluster name configMap
+		// TODO: remove this when NewAgentWithOperator is available in test-infra-definitions
+		_, err = corev1.NewConfigMap(ctx, "datadog-cluster-name", &corev1.ConfigMapArgs{
+			Metadata: metav1.ObjectMetaArgs{
+				Namespace: pulumi.String(common.NamespaceName),
+				Name:      pulumi.String("datadog-cluster-name"),
+			},
+			Data: pulumi.StringMap{
+				"DD_CLUSTER_NAME": pulumi.String(kindClusterName),
+			},
+		}, pulumi.Provider(kindKubeProvider))
+		if err != nil {
+			return err
+		}
+
+		env.FakeIntake = nil
+		env.Agent = nil
+
+		return nil
+
+	}, runner.ConfigMap{
+		"ddagent:deploy":                           auto.ConfigValue{Value: "false"},
+		"ddtestworkload:deploy":                    auto.ConfigValue{Value: "false"},
+		"ddagent:fakeintake":                       auto.ConfigValue{Value: "false"},
+		"dddogstatsd:deploy":                       auto.ConfigValue{Value: "false"},
+		"ddinfra:deployFakeintakeWithLoadBalancer": auto.ConfigValue{Value: "false"},
+		"ddagent:imagePullRegistry":                auto.ConfigValue{Value: "669783387624.dkr.ecr.us-east-1.amazonaws.com"},
+		"ddagent:imagePullUsername":                auto.ConfigValue{Value: "AWS"},
+		"ddagent:imagePullPassword":                auto.ConfigValue{Value: common.ImgPullPassword},
+		"ddinfra:kubernetesVersion":                auto.ConfigValue{Value: k8sVersion},
+	})
+}
