+++ test/e2e/tests/updaterSuite/rc-updater/api/client.go
