--- test/new-e2e/manifests/datadog-agent-minimum.yaml
@@ -1,19 +1,0 @@
-apiVersion: datadoghq.com/v2alpha1
-kind: DatadogAgent
-metadata:
-  name: datadog-minimum
-  namespace: e2e-operator
-  labels:
-    agent.datadoghq.com/e2e-test: datadog-agent-minimum
-spec:
-  global:
-    kubelet:
-      tlsVerify: false
-  features:
-    clusterChecks:
-      enabled: true
-      useClusterChecksRunners: true
-    logCollection:
-      enabled: true
-    liveContainerCollection:
-      enabled: true
